Move high order reducer to be within correct state slice
