+++ client/state/my-sites/reducer.js
@@ -3,7 +3,6 @@
  */
 import { combineReducers, withSchemaValidation } from 'calypso/state/utils';
 import sidebar from './sidebar/reducer';
-import { MY_SITES_SIDEBAR_SECTIONS_COLLAPSE_ALL } from 'calypso/state/action-types';
 
 const schema = {
 	type: 'object',
@@ -16,29 +15,6 @@ const schema = {
 	additionalProperties: false,
 };
 
-/**
- * Higher-order reducer to enable expanding/collapsing of all
- * the sidebar sections via a single action.
- *
- * @param reducer Function the keyed reducer to be enhanced
- */
-const withAllSectionsSidebarControls = ( reducer ) => ( state, action ) => {
-	switch ( action.type ) {
-		case MY_SITES_SIDEBAR_SECTIONS_COLLAPSE_ALL:
-			return Object.keys( state ).reduce( ( acc, curr ) => {
-				acc[ curr ] = {
-					...state[ curr ],
-					isOpen: false,
-				};
-
-				return acc;
-			}, {} );
-		default:
-			// Call original section-specific keyed reducer
-			return reducer( state, action );
-	}
-};
-
 export default combineReducers( {
-	sidebarSections: withSchemaValidation( schema, withAllSectionsSidebarControls( sidebar ) ),
+	sidebarSections: withSchemaValidation( schema, sidebar ),
 } );
